+++ src/main/java/fr/estecka/cmdbook/config/ConfigIO.java
@@ -1,4 +1,4 @@
-package tk.estecka.cmdbook.config;
+package fr.estecka.cmdbook.config;
 
 
 /*